+++ tests/helpers.py
@@ -1,15 +1,23 @@
 import asyncio
 import io
+import socket
 import typing
 
 
-class StatsdServer(asyncio.Protocol):
-    metrics: typing.List[bytes]
+class SupportsClose(typing.Protocol):
+    def close(self) -> None:
+        ...
 
-    def __init__(self):
-        self.service = None
+
+class StatsdServer(asyncio.DatagramProtocol, asyncio.Protocol):
+    metrics: typing.List[bytes]
+    server: typing.Union[SupportsClose, None]
+
+    def __init__(self, ip_protocol):
+        self.server = None
         self.host = '127.0.0.1'
         self.port = 0
+        self.ip_protocol = ip_protocol
         self.connections_made = 0
         self.connections_lost = 0
         self.message_counter = 0
@@ -26,25 +34,41 @@ class StatsdServer(asyncio.Protocol):
         await self._reset()
 
         loop = asyncio.get_running_loop()
-        self.service = await loop.create_server(lambda: self,
-                                                self.host,
-                                                self.port,
-                                                reuse_port=True)
-        listening_sock = self.service.sockets[0]
-        self.host, self.port = listening_sock.getsockname()
-        self.running.set()
-        try:
-            await self.service.serve_forever()
-            self.running.clear()
-        except asyncio.CancelledError:
-            self.close()
-            await self.service.wait_closed()
-        except Exception as error:
-            raise error
+        if self.ip_protocol == socket.IPPROTO_TCP:
+            server = await loop.create_server(lambda: self,
+                                              self.host,
+                                              self.port,
+                                              reuse_port=True)
+            self.server = server
+            listening_sock = server.sockets[0]
+            self.host, self.port = listening_sock.getsockname()
+            self.running.set()
+            try:
+                await server.serve_forever()
+            except asyncio.CancelledError:
+                self.close()
+                await server.wait_closed()
+            except Exception as error:
+                raise error
+            finally:
+                self.running.clear()
+
+        elif self.ip_protocol == socket.IPPROTO_UDP:
+            transport, protocol = await loop.create_datagram_endpoint(
+                lambda: self,
+                local_addr=(self.host, self.port),
+                reuse_port=True)
+            self.server = transport
+            self.host, self.port = transport.get_extra_info('sockname')
+            self.running.set()
+            try:
+                while not transport.is_closing():
+                    await asyncio.sleep(0.1)
+            finally:
+                self.running.clear()
 
     def close(self):
-        self.running.clear()
-        self.service.close()
+        self.server.close()
         for connected_client in self.transports:
             connected_client.close()
         self.transports.clear()
@@ -53,9 +77,6 @@ class StatsdServer(asyncio.Protocol):
         await self.running.wait()
 
     async def wait_closed(self):
-        if self.service.is_serving():
-            self.close()
-            await self.service.wait_closed()
         while self.running.is_set():
             await asyncio.sleep(0.1)
 
@@ -69,6 +90,13 @@ class StatsdServer(asyncio.Protocol):
 
     def data_received(self, data: bytes):
         self._buffer.write(data)
+        self._process_buffer()
+
+    def datagram_received(self, data: bytes, _addr):
+        self._buffer.write(data + b'\n')
+        self._process_buffer()
+
+    def _process_buffer(self):
         buf = self._buffer.getvalue()
         if b'\n' in buf:
             buf_complete = buf[-1] == ord('\n')