+++ CHANGELOG.rst
@@ -1,3 +1,3 @@
 Initial release
 ---------------
-- support for sending counters & timers to statsd over a TCP socket
+- support for sending counters & timers to statsd over a TCP or UDP socket